Move nestListInListItems to BlocksContentRenderer
